requirements.txt
Flask>=2.2.2
Pillow==8.1.2
tensorflow-cpu==2.11.0
scikit-learn
pandas